+++ examples/multi_thread/src/context.rs
@@ -25,7 +25,7 @@ pub struct Worker {
 }
 
 impl Agent for Worker {
-    type Reach = Context;
+    type Reach = Context<Self>;
     type Message = Msg;
     type Input = Request;
     type Output = Response;
