+++ examples/pub_sub/src/event_bus.rs
@@ -13,7 +13,7 @@ pub struct EventBus {
 }
 
 impl Agent for EventBus {
-    type Reach = Context;
+    type Reach = Context<Self>;
     type Message = ();
     type Input = Request;
     type Output = String;
